requirements.txt
fastapi
uvicorn
python-docx
Jinja2
python-multipart
mangum